refactor: jacocoで不要な処理などを削除
