+++ .circleci/config.yml
@@ -41,10 +41,6 @@ jobs:
           command: ./gradlew jacocoTestReport
 
       - run:
-          name: merge-jacoco
-          command: ./gradlew mergeJacoco
-
-      - run:
           name: run-danger-file
           command: bundle exec danger --verbose
