+++ build-logic/convention/src/main/kotlin/ksnd/hiraganaconverter/AndroidApplicationJacocoPugin.kt
@@ -1,6 +1,5 @@
 package ksnd.hiraganaconverter
 
-import kosenda.makecolor.configureJacoco
 import org.gradle.api.Plugin
 import org.gradle.api.Project
