+++ build-logic/convention/src/main/kotlin/ksnd/hiraganaconverter/AndroidApplicationPlugin.kt
@@ -1,7 +1,6 @@
 package ksnd.hiraganaconverter
 
 import com.android.build.api.dsl.ApplicationExtension
-import kosenda.makecolor.kotlinOptions
 import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
