+++ build-logic/convention/src/main/kotlin/ksnd/hiraganaconverter/Extensions.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor
+package ksnd.hiraganaconverter
 
 import com.android.build.api.dsl.CommonExtension
 import org.gradle.api.Project
@@ -7,11 +7,9 @@ import org.gradle.api.plugins.ExtensionAware
 import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.getByType
-import org.gradle.kotlin.dsl.register
 import org.gradle.kotlin.dsl.withType
 import org.gradle.testing.jacoco.plugins.JacocoPluginExtension
 import org.gradle.testing.jacoco.plugins.JacocoTaskExtension
-import org.gradle.testing.jacoco.tasks.JacocoReport
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
 
 /*
@@ -29,22 +27,6 @@ internal fun Project.configureJacoco() {
     configure<JacocoPluginExtension> {
         toolVersion = libs.findVersion("jacoco").get().toString()
     }
-    val jacocoTestReport = tasks.create("jacocoTestReport")
-    val testTaskName = "testProdDebugUnitTest"
-    val reportTask = tasks.register(
-        name = "jacoco${testTaskName}Report",
-        type = JacocoReport::class
-    ) {
-        dependsOn(testTaskName)
-        reports {
-            html.required.set(true)
-            xml.required.set(true)
-        }
-        classDirectories.setFrom(fileTree("${layout.buildDirectory.get()}/tmp/kotlin-classes/prodDebug"))
-        sourceDirectories.setFrom(files("$projectDir/src/main/java", "$projectDir/src/main/kotlin"))
-        executionData.from.add(fileTree("${layout.buildDirectory.get()}/jacoco/$testTaskName.exec"))
-    }
-    jacocoTestReport.dependsOn(reportTask)
     tasks.withType<Test>().configureEach {
         configure<JacocoTaskExtension> {
             isIncludeNoLocationClasses = true
